# Make initial auto readahead_size configurable